v3.2.0 - 2026-09-19

  35,862 Total lines  -  22,276 Code lines

Main changes:
  -> '--hide' takes each percentage by its own name now: 'files-percentages', 'lines-percentages',
      'code-percentages', 'comments-percentages', and 'change-percentages' for a '--diff'.
      'percentages' still takes all of them.
  -> mezura-core no longer forbids unsafe code: '#![forbid(unsafe_code)]' became '#![deny(unsafe_code)]',
      with four small unsafe blocks: two calls into the AVX2 routines (NEON on
      arm64), one reading a file that passed the ASCII check as text without a second validation,
      and one mapping a large file into memory. The mezura binary still forbids it.

Bugfixes:
  -> A string opener written inside a comment, such as Rust's 'r#"' or C++'s 'R"(', no longer
      swallows the rest of that line.
  -> The rows of the details table no longer spread apart because one language somewhere in it
      has nested rows.

Performance:
  -> Counting is measurably faster. Each file is now split into lines and scanned for its symbols in one pass
      over 64-byte blocks, where it used to take a search per line. On arm64, Apple Silicon included,
      that pass runs with NEON where it ran byte by byte. On Linux the walk no longer asks the size of
      every file, and files of 256 KB and up are mapped into memory (mmap).
  -> Release binaries are built under a profile of a real count. From source,
      'scripts/pgo-build.sh <a tree to count>' builds the same thing.

Other:
  -> Linux and macOS now start half as many producers and half as many consumers, since an open
      costs little there. Windows is unchanged. '--threads' still sets both by hand.
  -> MEZURA_PHASE_TIMING also reports when the first and the last counting thread finished.

+++ mezura-core/src/domain.rs
@@ -212,14 +212,14 @@ impl Language {
 
     // The whole width of one occurrence, which for a leveled pair depends on how many '=' it
     // carried; a plain or nesting pair ignores the level
-    pub(crate) fn comment_start_len(&self, symbol: u8, level: u8) -> usize {
+    pub(crate) fn comment_start_len(&self, symbol: u8, level: u32) -> usize {
         match self.get_comment_pair_of(symbol) {
             CommentPair::Leveled(pair) => pair.start_prefix.len() + level as usize + 1,
             CommentPair::Plain { start, .. } | CommentPair::Nesting { start, .. } => start.len()
         }
     }
 
-    pub(crate) fn comment_end_len(&self, symbol: u8, level: u8) -> usize {
+    pub(crate) fn comment_end_len(&self, symbol: u8, level: u32) -> usize {
         match self.get_comment_pair_of(symbol) {
             CommentPair::Leveled(pair) => pair.end_prefix.len() + level as usize + 1,
             CommentPair::Plain { end, .. } | CommentPair::Nesting { end, .. } => end.len()
